made duration a float to have more accurate results

+++ src/main/java/com/github/stickerifier/stickerify/media/MediaConstraints.java
@@ -17,7 +17,7 @@ public final class MediaConstraints {
 	static final long MAX_VIDEO_FILE_SIZE = 256_000L;
 	static final long MAX_ANIMATION_FILE_SIZE = 64_000L;
 	static final int MAX_ANIMATION_FRAMERATE = 60;
-	static final int MAX_ANIMATION_DURATION_SECONDS = 3;
+	static final float MAX_ANIMATION_DURATION_SECONDS = 3F;
 
 	private MediaConstraints() {
 		throw new UnsupportedOperationException();
